backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.7.1
requests==2.32.3
packaging==25.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.13
bcrypt==4.2.1
PyJWT==2.10.1
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
cellpose==3.1.1.1
